crawler/requirements.txt
playwright>=1.40.0
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.3.0
python-dotenv>=1.0.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
openai>=1.10.0
google-generativeai>=0.3.0
